Add missing cast for improper time_t use in computation.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@85611 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/md5/md5.c
@@ -162,7 +162,7 @@ MDTimeTrial(void)
 	 * Anderson, ric@Artisoft.COM.) */
 	printf
 	    ("Speed = %ld bytes/second\n",
-	    (long) TEST_BLOCK_LEN * (long) TEST_BLOCK_COUNT / ((endTime - startTime) != 0 ? (endTime - startTime) : 1));
+	    (long) TEST_BLOCK_LEN * (long) TEST_BLOCK_COUNT / ((endTime - startTime) != 0 ? (long)(endTime - startTime) : 1));
 }
 /*
  * Digests a reference suite of strings and prints the results.
